requirements.txt
flask==2.0.1
flask-sse==0.3.1
requests==2.25.1
beautifulsoup4==4.9.3
openai==0.27.0
replicate==0.3.0
pinecone-client==2.2.2
